抽取dao层
